Fixed publishing of DataPoint events
Now publishes points from all plots in cases where multiple plots have overlapping points.  Also only publishes if the point is different from the previously published point. Also publishes a null point when mousing off a point. All these changes are to bring it in line with the behavior of the BodyTrack grapher.