PlotContainer and SeriesPlotContainer now require the date axis to be passed in
This change helps by 1) eliminating the need for the getXAxis() method, which tried to figure out the date axis by looking at the contained plots. But, if there are no contained plots, getXAxis() returns null, even though there actually is a date axis associated with every plot container; and 2) no more chance of NPE's in methods which used to call getXAxis() since the X axis can no longer be null.